Some reformatting
